pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "impact"
version = "0.1.0"
description = "IMPACT — Standardized Model Development and Deployment Platform"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    "pandas>=2.0",
    "pydantic>=2.0",
    "PyYAML>=6.0",
    "pyarrow>=14.0",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
snowflake = [
    "snowflake-connector-python>=3.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.3",
]
all = [
    "impact[snowflake,dev]",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
